requirements.txt
camelot-py
pandas
numpy
dotenv
langchain
langchain_google_genai
langchain_core
langchain_community
mysql-connector-python
sqlalchemy
